+++ packages/elements/src/components/index.js
@@ -8,6 +8,7 @@ import TeamFairplay from './TeamFairplay.ce.vue';
 import TeamPenaltyKilling from './TeamPenaltyKilling.ce.vue';
 import TeamPowerplay from './TeamPowerplay.ce.vue';
 import TeamScoringEfficiency from './TeamScoringEfficiency.ce.vue';
+import PlayersNational from './PlayersNational.ce.vue';
 import FieldPlayersToi from './FieldPlayersToi.ce.vue';
 
 export {
@@ -21,5 +22,6 @@ export {
   TeamPenaltyKilling,
   TeamPowerplay,
   TeamScoringEfficiency,
+  PlayersNational,
   FieldPlayersToi,
 };