+++ packages/elements/src/index.js
@@ -11,6 +11,7 @@ import {
   TeamPenaltyKilling,
   TeamPowerplay,
   TeamScoringEfficiency,
+  PlayersNational,
   FieldPlayersToi,
 } from './components';
 
@@ -24,6 +25,7 @@ const TeamFairplayCE = defineCustomElement(TeamFairplay);
 const TeamPenaltyKillingCE = defineCustomElement(TeamPenaltyKilling);
 const TeamPowerplayCE = defineCustomElement(TeamPowerplay);
 const TeamScoringEfficiencyCE = defineCustomElement(TeamScoringEfficiency);
+const PlayersNationalCE = defineCustomElement(PlayersNational);
 const FieldPlayersToiCE = defineCustomElement(FieldPlayersToi);
 
 export {
@@ -37,6 +39,7 @@ export {
   TeamPenaltyKilling,
   TeamPowerplay,
   TeamScoringEfficiency,
+  PlayersNational,
   StandingsCE,
   ScheduleCE,
   FieldPlayersLeaderCE,
@@ -47,6 +50,7 @@ export {
   TeamPenaltyKillingCE,
   TeamPowerplayCE,
   TeamScoringEfficiencyCE,
+  PlayersNationalCE,
   FieldPlayersToiCE,
 };
 
@@ -61,6 +65,7 @@ export function register() {
   customElements.define('mjsz-vbr-team-penalty-killing', TeamPenaltyKillingCE);
   customElements.define('mjsz-vbr-team-powerplay', TeamPowerplayCE);
   customElements.define('mjsz-vbr-team-scoring', TeamScoringEfficiencyCE);
+  customElements.define('mjsz-vbr-players-national', PlayersNationalCE);
   customElements.define('mjsz-vbr-players-toi', FieldPlayersToiCE);
 }