+++ packages/gamecenter/src/components/game/GameData.vue
@@ -46,6 +46,10 @@ const { t } = useI18n();
           <IconSheet class="is-icon" />
           {{ t('sheet') }}
         </a>
+        <a v-if="gameData.toiReportUrl" :href="gameData.toiReportUrl" target="_blank">
+          <IconSheet class="is-icon" />
+          {{ t('toiSheet') }}
+        </a>
         <a v-if="gameData.video" :href="gameData.video" target="_blank">
           <IconYoutube class="is-icon" />
           {{ t('video') }}